techdocs-common: avoid global type declarations
