+++ packages/techdocs-common/src/stages/publish/awsS3.test.ts
@@ -26,7 +26,7 @@ import { AwsS3Publish } from './awsS3';
 
 // NOTE: /packages/techdocs-common/__mocks__ is being used to mock aws-sdk client library
 
-const rootDir = global.rootDir;
+const rootDir = (global as any).rootDir; // Set by setupTests.ts
 
 const getEntityRootDir = (entity: Entity) => {
   const {
