+++ packages/techdocs-common/src/testUtils/types.ts
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-declare module NodeJS {
-  interface Global {
-    rootDir: string;
-    storageFilesMock: IStorageFilesMock;
-  }
+export interface IStorageFilesMock {
+  emptyFiles(): void;
+  fileExists(targetPath: string): boolean;
+  readFile(targetPath: string): Buffer;
+  writeFile(targetPath: string, sourcePath: string): void;
+  writeFile(targetPath: string, sourceBuffer: Buffer): void;
+  writeFile(targetPath: string, source: string | Buffer): void;
 }
