--- packages/techdocs-common/src/types.d.ts
@@ -1,24 +1,0 @@
-/*
- * Copyright 2020 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-interface IStorageFilesMock {
-  emptyFiles(): void;
-  fileExists(targetPath: string): boolean;
-  readFile(targetPath: string): Buffer;
-  writeFile(targetPath: string, sourcePath: string): void;
-  writeFile(targetPath: string, sourceBuffer: Buffer): void;
-  writeFile(targetPath: string, source: string | Buffer): void;
-}
